feat: criado método format response em github client
