+++ src/client/github.client.ts
@@ -1,4 +1,6 @@
+import { IGithubRepositoryInfo } from '@src/interfaces/IGithubRepository'
 import axios, { AxiosResponse } from 'axios'
+import _ from 'lodash'
 
 export class GitHubClient {
   constructor(private request = axios) {}
@@ -8,7 +10,7 @@ export class GitHubClient {
       const response = await this.request.get(
         `https://api.github.com/search/repositories?q=language:${language}&sort=stars&order=desc&per_page=5`
       )
-      return response.data.items
+      return response
     } catch (error) {
       if (!error.status && !error.response) {
         throw new Error(
@@ -19,5 +21,30 @@ export class GitHubClient {
         `Unexpected error returned by the GitHub API: ${error.response.data.message}`
       )
     }
+  }
+
+  public async formatResponse(
+    languages: string[]
+  ): Promise<IGithubRepositoryInfo[]> {
+    const arrayRepos: IGithubRepositoryInfo[] = await Promise.all(
+      languages.map(async (lang: string) => {
+        const response = await this.fetchRepositories(lang)
+        return response.data.items
+      })
+    )
+
+    const response = _.flattenDeep(arrayRepos).map((element) => {
+      return {
+        id: element.id,
+        name: element.name,
+        full_name: element.full_name,
+        description: element.description,
+        url: element.url,
+        stargazers_count: element.stargazers_count,
+        language: element.language,
+      }
+    })
+
+    return response
   }
 }
