+++ src/controllers/repo.controller.ts
@@ -2,8 +2,6 @@ import { User } from '@src/models/user.model'
 import { Request, Response } from 'express'
 import { getConnection } from 'typeorm'
 import { GitHubClient } from '@src/client/github.client'
-import { IGithubRepository } from '@src/interfaces/IGithubRepository'
-import _ from 'lodash'
 
 const client = new GitHubClient()
 
@@ -12,6 +10,10 @@ export async function findTopRepositories(
   res: Response
 ): Promise<void> {
   try {
+    if (!req.body.languages) {
+      res.status(422).json({ code: 422, message: 'Validation Failed' })
+    }
+
     const { languages } = req.body
     const result = await getConnection()
       .createQueryBuilder()
@@ -25,28 +27,10 @@ export async function findTopRepositories(
       return
     }
 
-    const arrayRepos: IGithubRepository[] = await Promise.all(
-      languages.map(async (lang: string) => {
-        const response = await client.fetchRepositories(lang)
-        return response
-      })
-    )
-
-    const response = _.flattenDeep(arrayRepos).map((element) => {
-      return {
-        id: element.id,
-        name: element.name,
-        full_name: element.full_name,
-        description: element.description,
-        url: element.url,
-        stargazers_count: element.stargazers_count,
-        language: element.language,
-      }
-    })
+    const response = await client.formatResponse(languages)
 
     res.send(response)
   } catch (error) {
     res.status(500).send({ code: 500, message: error.message })
-    return
   }
 }
